% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-summaries.R
\name{generate_summary_table}
\alias{generate_summary_table}
\title{Generate highlight summary table}
\usage{
generate_summary_table(use_20 = TRUE)
}
\arguments{
\item{use_20}{Wether to use 20% or 100% of assumed daily nutrient intake}
}
\value{
A interactive table
}
\description{
Generate a summary table of nutritional scenario in Timor-Est by region.
}
