% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_model_data}
\alias{get_model_data}
\title{Generate Modeling Datasets}
\usage{
get_model_data()
}
\value{
A list containing the four datasets. The list also includes kmeans profiles plots.
}
\description{
Prepares and returns datasets for use with an ML model. This function generates
four datasets: two focus exclusively on gill net data, and the other two cover all other gear types.
These datasets are specific to Atauro and other municipalities.
}
\details{
The decision to divide the dataset into four subsets is driven by the
imbalance present in the raw survey data from Peskas. A significant portion of the data (>40%)
relates to Atauro, and the most common gear type is the gill net.

By splitting the data into these subsets, we aim to reduce the bias caused by
the overrepresentation of Atauro and gill net observations. Additionally, analyzing
only the gill net data separately allows for an evaluation of the impact of mesh size
on predicting nutrient profiles.
}
\examples{
\dontrun{
data_list <- get_model_data()

structured_results <-
  purrr::imap(
    data_list, ~ run_xgmodel(
      dataframe = .x$dataframe, step_other = .x$step_other
    )
  ) \%>\%
  setNames(paste0("model_", names(.)))
}

}
