% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb-model.R
\name{get_shaps}
\alias{get_shaps}
\title{Extract SHAP Values for Specific Features}
\usage{
get_shaps(shap_object = NULL, model_type = NULL)
}
\arguments{
\item{shap_object}{A list containing the SHAP values and corresponding feature values.
Must have elements `X` and `S`, where `X` contains feature values and
`S` contains corresponding SHAP values.}
}
\value{
A `tibble` containing columns for mesh size (`mesh_fact`), habitat (`habitat_fact`),
        and their corresponding SHAP values (`mesh_shap`).
}
\description{
This function extracts specific features and their corresponding SHAP values from a SHAP object.
It is designed to work with SHAP objects that contain features such as mesh size and habitat.
}
\details{
The function specifically extracts 'mesh size' and 'habitat' features from the `X` component
and their corresponding SHAP values from the `S` component of the `shap_object`.
It is used as a utility function within other functions handling SHAP values.
}
\examples{
\dontrun{
shap_data <- list(
  X = data.frame(mesh_size = 1:10, habitat = 11:20),
  S = matrix(runif(30), ncol = 3)
)
extracted_shaps <- get_shaps(shap_data)
}
}
