% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb-model.R
\name{plot_model_shaps}
\alias{plot_model_shaps}
\title{Plot SHAP Values for Multiple Models}
\usage{
plot_model_shaps(
  data_shaps = NULL,
  model_type = NULL,
  alpha = 0.2,
  cols = 1,
  drop_legend = FALSE
)
}
\arguments{
\item{data_shaps}{A data frame or list of data frames containing SHAP values and corresponding features.}

\item{model_type}{A character string specifying the model type, passed to the `plot_shap` function.}

\item{alpha}{The alpha value for geom_jitter in ggplot2, controlling point transparency.}

\item{cols}{The number of columns in the plot layout.}

\item{drop_legend}{Wether to return legend. Default is TRUE.}
}
\value{
A ggplot object representing the combined grid of SHAP value plots for different models.
}
\description{
This function visualizes SHAP values across multiple models or model types, integrating the `plot_shap` function.
}
\details{
The function uses the `shapviz` package for initial processing and then applies `plot_shap` to each model.
It arranges the resulting plots in a grid format, optionally including a legend and common axes labels.
}
\examples{
\dontrun{
plot_model_shaps(data_shaps = my_model_shaps, model_type = "gn", alpha = 0.2, cols = 2)
}
}
