% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb-model.R
\name{plot_shap}
\alias{plot_shap}
\title{Plot SHAP Values for Different Model Types}
\usage{
plot_shap(shap_object = NULL, model_type = NULL, alpha = NULL)
}
\arguments{
\item{shap_object}{A list containing the SHAP values and corresponding feature values. Must have elements `X` and `S`.}

\item{model_type}{A character string specifying the model type. For this function, it can be "gn" or other types.}

\item{alpha}{The alpha value for the geom_jitter layer in ggplot2, controlling the transparency of points.}
}
\value{
A ggplot object representing the SHAP value scatter plot.
}
\description{
This function generates a scatter plot of SHAP (SHapley Additive exPlanations) values for different model types.
It is specifically designed for either "gn" or other model types, plotting different aspects of the data.
}
\details{
If `model_type` is "gn", the function plots `mesh_size`, `habitat`, and their corresponding SHAP values.
Otherwise, it plots `habitat_gear`, `vessel_type`, and their corresponding SHAP values.
}
\examples{
\dontrun{
plot_shap(shap_object = my_shap_data, model_type = "gn", alpha = 0.5)
}
}
