% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-df.R
\name{rename_nutrients_mu}
\alias{rename_nutrients_mu}
\title{Rename nutrients columns}
\usage{
rename_nutrients_mu(df = NULL, hyphen = FALSE)
}
\arguments{
\item{df}{Dataframe containing nutrients columns to rename.}

\item{hyphen}{Whether to use hyphen in renaming "omega_3" and "vitamin_a".}
}
\value{
A dataframe with renamed columns
}
\description{
Rename nutrients column with the format "Nutrient_mu" (from GitHub raw models data)
with "nutrient".
}
\examples{
\dontrun{
rename_nutrients_mu(nutrients_table)
}
}
