% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb-model.R
\name{run_kernelshap}
\alias{run_kernelshap}
\title{Run KernelSHAP on a Model}
\usage{
run_kernelshap(model)
}
\arguments{
\item{model}{A model object that contains at least two elements: `$train_data`, a dataframe of training data,
and `$fit`, the fitted model object.}
}
\value{
Returns an object of class `kernelshap`, which contains the SHAP (SHapley Additive exPlanations) values computed
for each feature in the training data, explaining the model output.
}
\description{
This function applies the KernelSHAP algorithm to a given model. It uses a subset of the training data
to explain the predictions made by the model.
}
\details{
The function first removes the last column from the training data. It then selects a random subset of 500 observations
from this modified training dataset to serve as the background dataset for KernelSHAP. The KernelSHAP explanation
is then computed on the entire (modified) training dataset with these settings.
}
\examples{
\dontrun{
# Assuming `model` is a pre-trained model with the required structure
shap_values <- run_kernelshap(model)
}
}
