% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb-model.R
\name{run_xgmodel}
\alias{run_xgmodel}
\title{Run XGModel}
\usage{
run_xgmodel(dataframe = NULL, step_other = NULL, n_cores = NULL)
}
\arguments{
\item{dataframe}{A dataframe to be used for modeling. This should contain
both the features and the target variable.}

\item{step_other}{Parameters for `recipes::step_other` function to handle
factor levels with few occurrences in the dataset.}

\item{n_cores}{Number of cores used for parallel processing in `doParallel::registerDoParallel`.}
}
\value{
A list containing various elements including the final model fit,
variable importance plot, ROC curves, and AUC value.
}
\description{
This function facilitates the execution of an XGBoost model using a given
dataset. It includes processes like data splitting, cross-validation,
preparation of machine learning recipe, hyperparameter tuning, and model evaluation.
The function returns a list containing the final model fit,
variable importance plot, ROC curves, and AUC value.
}
\examples{
\dontrun{
data(iris)
results <- run_xgmodel(
  dataframe = iris,
  step_other = "Other",
  n_cores = 6
)
print(results)
}
}
